+++ src/components/FormField/style.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/base';
+@import '../../styles/base';
 
 $labelWidth: 10rem;
